Update package registry in package.json and enhance npm-publish-github-packages workflow
This commit changes the npm registry in package.json from GitHub Packages to the public npm registry. Additionally, it updates the npm-publish-github-packages workflow to include a name for the publish step and sets the NPM_CONFIG_PROVENANCE environment variable to true, improving the clarity and functionality of the publishing process.